+++ src/AspNetCore.EnumMemberNameBinding.OpenApi/EnumMemberNameOpenApiOptionsExtensions.cs
@@ -1,3 +1,6 @@
+using System.Diagnostics.CodeAnalysis;
+
+using AspNetCore.EnumMemberNameBinding;
 using AspNetCore.EnumMemberNameBinding.OpenApi;
 
 namespace Microsoft.AspNetCore.OpenApi;
@@ -18,6 +21,8 @@ public static class EnumMemberNameOpenApiOptionsExtensions {
     /// builder.Services.AddOpenApi(options => options.AddEnumMemberNames());
     /// </code>
     /// </example>
+    [RequiresUnreferencedCode(EnumMemberNames.ReflectionRequirement)]
+    [RequiresDynamicCode(EnumMemberNames.ReflectionRequirement)]
     public static OpenApiOptions AddEnumMemberNames(this OpenApiOptions options) {
         ArgumentNullException.ThrowIfNull(options);